Remove retires from ModEasy
