+++ osu.Game/Rulesets/Mods/ModEasy.cs
@@ -2,17 +2,13 @@
 // See the LICENCE file in the repository root for full licence text.
 
 using System;
-using Humanizer;
-using osu.Framework.Bindables;
 using osu.Framework.Graphics.Sprites;
 using osu.Game.Beatmaps;
-using osu.Game.Configuration;
 using osu.Game.Graphics;
-using osu.Game.Rulesets.Scoring;
 
 namespace osu.Game.Rulesets.Mods
 {
-    public abstract class ModEasy : Mod, IApplicableToDifficulty, IApplicableFailOverride, IApplicableToHealthProcessor
+    public abstract class ModEasy : Mod, IApplicableToDifficulty
     {
         public override string Name => "Easy";
         public override string Acronym => "EZ";
@@ -22,49 +18,17 @@ namespace osu.Game.Rulesets.Mods
         public override bool Ranked => true;
         public override Type[] IncompatibleMods => new[] { typeof(ModHardRock), typeof(ModDifficultyAdjust) };
 
-        [SettingSource("Extra Lives", "Number of extra lives")]
-        public Bindable<int> Retries { get; } = new BindableInt(2)
-        {
-            MinValue = 0,
-            MaxValue = 10
-        };
-
-        public override string SettingDescription => Retries.IsDefault ? string.Empty : $"{"lives".ToQuantity(Retries.Value)}";
-
-        private int retries;
-
-        private BindableNumber<double> health;
-
-        public void ReadFromDifficulty(BeatmapDifficulty difficulty)
+        public virtual void ReadFromDifficulty(BeatmapDifficulty difficulty)
         {
         }
 
-        public void ApplyToDifficulty(BeatmapDifficulty difficulty)
+        public virtual void ApplyToDifficulty(BeatmapDifficulty difficulty)
         {
             const float ratio = 0.5f;
             difficulty.CircleSize *= ratio;
             difficulty.ApproachRate *= ratio;
             difficulty.DrainRate *= ratio;
-            difficulty.OverallDifficulty *= ratio;
-
-            retries = Retries.Value;
-        }
-
-        public bool PerformFail()
-        {
-            if (retries == 0) return true;
-
-            health.Value = health.MaxValue;
-            retries--;
-
-            return false;
-        }
-
-        public bool RestartOnFail => false;
-
-        public void ApplyToHealthProcessor(HealthProcessor healthProcessor)
-        {
-            health = healthProcessor.Health.GetBoundCopy();
+            difficulty.OverallDifficulty *= ratio;;
         }
     }
 }
